sfm: drop _glomap suffix + sfm_ext namespace from remaining fork files
Three-stage cleanup mirroring the TrackEngine fold (commit 77c0574b).
Net result: ZERO _glomap files in colmap/sfm/, ZERO sfm_ext namespace
references. All fork-extension code now lives in plain colmap::
namespace alongside upstream code, with no naming tax.

Stage 1 - two_view_geometry_glomap deletion (-79 LOC):
- 3 helpers (CheckCheirality single-pair, GetOrientationSignum,
  SampsonError Vec3 depth-aware) had ONE caller each
  (image_pair_inliers). Folded as anonymous-namespace statics in the
  consumer file. Deleted two_view_geometry_glomap.{h,cc} +
  CMakeLists entry.

Stage 2 - track_filter_glomap rename (no LOC delta, +/- ~250):
- Renamed src/colmap/sfm/track_filter_glomap.{h,cc} -> track_filter.{h,cc}.
- Dropped sfm_ext:: namespace -> plain colmap::.
- Converted struct TrackFilter::FilterX static methods to free
  functions FilterTracksByAngle / FilterTrackTriangulationAngle.
- Pycolmap binding: pycolmap.sfm_ext.filter_tracks_by_angle ->
  pycolmap.filter_tracks_by_angle (top-level). Same for
  filter_track_triangulation_angle. Test file + downstream callers
  updated.

Stage 3 - image_pair_inliers_glomap rename (no LOC delta):
- Renamed src/colmap/sfm/image_pair_inliers_glomap.{h,cc} ->
  image_pair_inliers.{h,cc}.
- Dropped sfm_ext:: namespace -> plain colmap::.
- Behavior preserved verbatim (hard-coded thres_epipole_nodepth,
  disabled Gate-3 angle check, F-branch positive_count==negative_count
  silent drop). TODOs for later cleanup left in-place.
- Pycolmap binding: pycolmap.image_pairs_inlier_count was already
  top-level; renamed BindImagePairInliersGlomap -> BindImagePairInliers.
- Renamed src/pycolmap/sfm/track_establishment_glomap.cc ->
  track_establishment.cc; renamed BindTrackEstablishmentGlomap ->
  BindTrackEstablishment. Same fold treatment.

Pycolmap top-level surface (no sfm_ext submodule any more):
- pycolmap.establish_full_tracks / find_tracks_for_problem
- pycolmap.filter_tracks_by_angle / filter_track_triangulation_angle
- pycolmap.image_pairs_inlier_count
- pycolmap.TrackEstablishmentOptions / TrackSubsampleOptions / InlierThresholdOptions

Removed back-compat alias pycolmap.glomap_ra = pycolmap.sfm_ext (no
remaining callers; sfm_ext submodule itself is gone).

Verified: scripts/build.sh clean. 7 ctest groups pass (scene/track,
scene/image, sfm/track_establishment, sfm/track_filter,
sfm/image_pair_inliers, estimators/global_positioning, estimators/
rotation_averaging) - 100% in 0.29s.
